Create WordCloud.py
write a python script that will prompt the user to input the name of a text file to be processed. 
The program should read the through the contents of the file where it will find a series of comma separated text strings.  Each text string should have the spaces removed from the string and it should be written to a new text file called OUTPUT.txt  
Each comma separated string should appear as a new line in the output text file.
The file may contain special characters such as &, ? and - which should be left unchanged.